+++ README.md
@@ -133,6 +133,14 @@
    - Загрузите файлы
    - В настройках репозитория включите GitHub Pages
    - Выберите ветку `main` и папку `/ (root)`
+### Сборка
+
+Исходный код находится в каталоге `src`. Для создания `script.js` выполните:
+
+```bash
+node build.js
+```
+
 
 ## 📁 Структура проекта
