+++ src/data/budgetData.js
@@ -1,0 +1,40 @@
+const budgetData = [
+  {
+    category: "Аренда кемпера",
+    amount: "18 000–30 000 NOK (≈ 1 570–2 620 €)",
+    note: "7 дней, средний кемпер 2 200–3 200 NOK/сутки с страховкой",
+  },
+  {
+    category: "Топливо и дороги",
+    amount: "5 000–7 000 NOK (≈ 440–610 €)",
+    note: "≈200 л дизеля по 19,7 NOK + AutoPASS/платные дороги",
+  },
+  {
+    category: "Кемпинги",
+    amount: "2 000–3 500 NOK (≈ 175–305 €)",
+    note: "6 ночей, 250–350 NOK/ночь с электричеством",
+  },
+  {
+    category: "Паромы",
+    amount: "1 000–2 000 NOK (≈ 90–175 €)",
+    note: "Eidsdal-Linge, Åfarnes-Sølsnes, Molde-Vestnes, Geiranger и др.",
+  },
+  {
+    category: "Еда",
+    amount: "6 000–10 000 NOK (≈ 525–875 €)",
+    note: "Продукты + 2 ресторана в Ålesund/Loen",
+  },
+  {
+    category: "Развлечения",
+    amount: "3 000–5 000 NOK (≈ 265–435 €)",
+    note: "Via Ferrata Loen, Romsdalsgondolen, Dalsnibba",
+  },
+  {
+    category: "Прочее",
+    amount: "2 000–5 000 NOK (≈ 175–435 €)",
+    note: "Сувениры, душевые жетоны, мелкие расходы",
+  },
+];
+
+
+if (typeof module !== "undefined") { module.exports = { budgetData }; }
