+++ src/data/checklistData.js
@@ -1,0 +1,41 @@
+const checklistData = {
+  documents: [
+    "Загранпаспорта (всех участников)",
+    "Водительские права (международные + национальные)",
+    "Страховка медицинская",
+    "Банковские карты (минимум 2 разных)",
+    "Наличные NOK (~3000-5000 крон)",
+    "Копии всех документов",
+    "Подтверждения бронирований",
+  ],
+  clothing: [
+    "Термобелье (2 комплекта)",
+    "Мембранная куртка и штаны",
+    "Дождевик",
+    "Пуховка или теплая куртка",
+    "Треккинговые ботинки (разношенные!)",
+    "Шапка и перчатки (даже летом)",
+    "Быстросохнущие штаны (2-3 пары)",
+    "Купальник/плавки",
+  ],
+  gear: [
+    "Перчатки для via ferrata (усиленная ладонь)",
+    "Налобный фонарь + батарейки",
+    "Бинокль 8x или 10x (для тупиков)",
+    "Треккинговые палки",
+    "Рюкзак дневной 20-30л",
+    "Солнцезащитный крем SPF 30+",
+    "Репеллент от комаров",
+  ],
+  camper: [
+    "Переходники для розеток",
+    "Удлинитель 10-25м",
+    "Веревка для белья + прищепки",
+    "Мусорные пакеты",
+    "Туалетная химия",
+    "Термос",
+    "Контейнеры для еды",
+  ],
+};
+
+if (typeof module !== "undefined") { module.exports = { checklistData }; }
