+++ src/data/tripData.js
@@ -456,459 +456,3 @@ const tripData = {
 };
 
 if (typeof module !== "undefined") { module.exports = { tripData }; }
-
-
-const checklistData = {
-  documents: [
-    "Загранпаспорта (всех участников)",
-    "Водительские права (международные + национальные)",
-    "Страховка медицинская",
-    "Банковские карты (минимум 2 разных)",
-    "Наличные NOK (~3000-5000 крон)",
-    "Копии всех документов",
-    "Подтверждения бронирований",
-  ],
-  clothing: [
-    "Термобелье (2 комплекта)",
-    "Мембранная куртка и штаны",
-    "Дождевик",
-    "Пуховка или теплая куртка",
-    "Треккинговые ботинки (разношенные!)",
-    "Шапка и перчатки (даже летом)",
-    "Быстросохнущие штаны (2-3 пары)",
-    "Купальник/плавки",
-  ],
-  gear: [
-    "Перчатки для via ferrata (усиленная ладонь)",
-    "Налобный фонарь + батарейки",
-    "Бинокль 8x или 10x (для тупиков)",
-    "Треккинговые палки",
-    "Рюкзак дневной 20-30л",
-    "Солнцезащитный крем SPF 30+",
-    "Репеллент от комаров",
-  ],
-  camper: [
-    "Переходники для розеток",
-    "Удлинитель 10-25м",
-    "Веревка для белья + прищепки",
-    "Мусорные пакеты",
-    "Туалетная химия",
-    "Термос",
-    "Контейнеры для еды",
-  ],
-};
-
-if (typeof module !== "undefined") { module.exports = { checklistData }; }
-
-
-const budgetData = [
-  {
-    category: "Аренда кемпера",
-    amount: "18 000–30 000 NOK (≈ 1 570–2 620 €)",
-    note: "7 дней, средний кемпер 2 200–3 200 NOK/сутки с страховкой",
-  },
-  {
-    category: "Топливо и дороги",
-    amount: "5 000–7 000 NOK (≈ 440–610 €)",
-    note: "≈200 л дизеля по 19,7 NOK + AutoPASS/платные дороги",
-  },
-  {
-    category: "Кемпинги",
-    amount: "2 000–3 500 NOK (≈ 175–305 €)",
-    note: "6 ночей, 250–350 NOK/ночь с электричеством",
-  },
-  {
-    category: "Паромы",
-    amount: "1 000–2 000 NOK (≈ 90–175 €)",
-    note: "Eidsdal-Linge, Åfarnes-Sølsnes, Molde-Vestnes, Geiranger и др.",
-  },
-  {
-    category: "Еда",
-    amount: "6 000–10 000 NOK (≈ 525–875 €)",
-    note: "Продукты + 2 ресторана в Ålesund/Loen",
-  },
-  {
-    category: "Развлечения",
-    amount: "3 000–5 000 NOK (≈ 265–435 €)",
-    note: "Via Ferrata Loen, Romsdalsgondolen, Dalsnibba",
-  },
-  {
-    category: "Прочее",
-    amount: "2 000–5 000 NOK (≈ 175–435 €)",
-    note: "Сувениры, душевые жетоны, мелкие расходы",
-  },
-];
-
-
-if (typeof module !== "undefined") { module.exports = { budgetData }; }
-
-
-const defaultNotes = [
-  {
-    text: "Проверить прогноз погоды для Dalsnibba перед подъемом",
-    author: "Система",
-    timestamp: new Date(),
-  },
-  {
-    text: "Забронировать Via Ferrata Loen как можно скорее",
-    author: "Система",
-    timestamp: new Date(),
-  },
-];
-
-const StorageService = {
-  loadNotes() {
-    return JSON.parse(localStorage.getItem("norwayTripNotes")) || [...defaultNotes];
-  },
-  saveNotes(notes) {
-    localStorage.setItem("norwayTripNotes", JSON.stringify(notes));
-  },
-  loadChecklist() {
-    return JSON.parse(localStorage.getItem("norwayTripChecklist")) || {};
-  },
-  saveChecklist(data) {
-    localStorage.setItem("norwayTripChecklist", JSON.stringify(data));
-  },
-  loadCurrentSection() {
-    return localStorage.getItem("norwayTripCurrentSection") || "itinerary";
-  },
-  saveCurrentSection(section) {
-    localStorage.setItem("norwayTripCurrentSection", section);
-  },
-};
-
-if (typeof module !== 'undefined') { module.exports = { StorageService, defaultNotes }; }
-
-
-const NavigationService = {
-  showSection(sectionName) {
-    document.querySelectorAll('.section').forEach(sec => sec.classList.add('hidden'));
-    document.getElementById(sectionName).classList.remove('hidden');
-    document.querySelectorAll('.nav-btn').forEach(btn => btn.classList.remove('active'));
-    event.target && event.target.classList.add('active');
-    localStorage.setItem('norwayTripCurrentSection', sectionName);
-  },
-};
-
-if (typeof module !== 'undefined') { module.exports = { NavigationService }; }
-
-
-class Modal {
-  constructor() {
-    this.modal = document.getElementById('location-modal');
-    this.titleEl = document.getElementById('modal-title');
-    this.contentEl = document.getElementById('modal-content');
-  }
-
-  showLocationDetails(day, location, tripData) {
-    const dayData = tripData.days.find(d => d.day == day);
-    const loc = dayData.locations.find(l => l.name === location);
-    this.titleEl.textContent = loc.name;
-    this.contentEl.innerHTML = `
-    <div class="space-y-4">
-      <div class="flex items-center space-x-2">
-        <span class="text-lg">${getLocationTypeEmoji(loc.type)}</span>
-        <span class="text-muted">${loc.time}</span>
-      </div>
-      <p class="text-primary">${loc.description}</p>
-      ${loc.notes ? `<div class="bg-amber-50 border border-amber-200 rounded-lg p-3">
-          <h4 class="font-semibold text-amber-800 mb-1">💡 Важные заметки:</h4>
-          <p class="text-amber-700">${loc.notes}</p>
-        </div>` : ''}
-      <div class="flex space-x-2">
-        ${loc.maps ? `<a href="${loc.maps}" target="_blank" class="btn">🗺️ Открыть на карте</a>` : ''}
-        <button onclick="notesComponent.addLocationNote('${day}', '${location}')" class="btn-secondary">💭 Добавить заметку</button>
-      </div>
-    </div>`;
-    this.modal.style.display = 'flex';
-  }
-
-  close() {
-    this.modal.style.display = 'none';
-  }
-}
-
-if (typeof module !== 'undefined') { module.exports = { Modal }; }
-
-
-class Timeline {
-  constructor(tripData, modal) {
-    this.tripData = tripData;
-    this.modal = modal;
-    this.container = document.querySelector('#itinerary .grid');
-  }
-
-  render() {
-    const container = this.container;
-    container.className = 'timeline';
-    container.innerHTML = '';
-    this.tripData.days.forEach(day => {
-      const timelineItem = document.createElement('div');
-      timelineItem.className = 'timeline-item';
-      timelineItem.innerHTML = `
-      <div class="timeline-dot status-${day.status}">${day.day}</div>
-      <div class="timeline-card">
-        <div class="timeline-header">
-          <div class="timeline-title">${day.title}</div>
-          <div class="timeline-meta"><span>${day.date}</span><span>${day.locations.length} остановок</span></div>
-          <div class="timeline-stats"><span>📍 ${day.distance}</span><span>⏱️ ${day.time}</span></div>
-        </div>
-      </div>
-      <div class="timeline-locations">
-        <div class="locations-header">📋 Детали дня</div>
-        ${day.locations.map(loc => `
-          <div class="location-item ${loc.type === 'hidden_gem' ? 'hidden-gem' : ''}" onclick="modalComponent.showLocationDetails('${day.day}','${loc.name}', tripData)">
-            <div class="location-icon ${loc.type}">${getLocationTypeEmoji(loc.type)}</div>
-            <div class="location-details">
-              <div class="location-name">${loc.name}</div>
-              <div class="location-time">${loc.time}</div>
-              <div class="location-description">${loc.description}</div>
-              ${loc.notes ? `<div class="location-notes">💡 ${loc.notes}</div>` : ''}
-            </div>
-          </div>`).join('')}
-      </div>`;
-      container.appendChild(timelineItem);
-    });
-  }
-
-  renderHiddenGems() {
-    const container = document.getElementById('hidden-gems-container');
-    container.innerHTML = '';
-    this.tripData.days.forEach(day => {
-      const gems = day.locations.filter(loc => loc.type === 'hidden_gem');
-      if (gems.length) {
-        const daySection = document.createElement('div');
-        daySection.className = 'bg-white border border-gray-200 rounded-lg p-6';
-        daySection.innerHTML = `
-        <h3 class="text-lg font-semibold text-primary mb-4 flex items-center">
-          <span class="bg-purple-100 text-purple-800 rounded-full w-8 h-8 flex items-center justify-center text-sm font-bold mr-3">${day.day}</span>
-          День ${day.day}: ${day.title}
-        </h3>
-        <div class="space-y-4">
-          ${gems.map(gem => `
-            <div class="location-item hidden-gem cursor-pointer" onclick="modalComponent.showLocationDetails('${day.day}','${gem.name}', tripData)">
-              <div class="location-icon ${gem.type}">${getLocationTypeEmoji(gem.type)}</div>
-              <div class="location-details">
-                <div class="location-name">${gem.name}</div>
-                <div class="location-time">${gem.time}</div>
-                <div class="location-description">${gem.description}</div>
-                ${gem.notes ? `<div class="location-notes">💡 ${gem.notes}</div>` : ''}
-              </div>
-            </div>`).join('')}
-        </div>`;
-        container.appendChild(daySection);
-      }
-    });
-    const total = this.tripData.days.reduce((t, d) => t + d.locations.filter(l => l.type === 'hidden_gem').length,0);
-    if (total) {
-      const summary = document.createElement('div');
-      summary.className = 'bg-gradient-to-r from-purple-50 to-pink-50 border border-purple-200 rounded-lg p-6 text-center';
-      summary.innerHTML = `<h3 class="text-lg font-semibold text-purple-800 mb-2">✨ Всего найдено ${total} тайных мест</h3><p class="text-purple-600 text-sm">Эти места помогут вам погрузиться в аутентичную норвежскую культуру и избежать туристических толп</p>`;
-      container.appendChild(summary);
-    }
-  }
-}
-
-if (typeof module !== 'undefined') { module.exports = { Timeline }; }
-
-
-class Checklist {
-  constructor(data, progress) {
-    this.data = data;
-    this.progress = progress;
-    this.container = document.querySelector('#checklist .grid');
-  }
-
-  render() {
-    this.container.innerHTML = '';
-    Object.entries(this.data).forEach(([category, items]) => {
-      const div = document.createElement('div');
-      div.className = 'bg-white border border-gray-200 rounded-lg p-4';
-      div.innerHTML = `
-        <h3 class="font-semibold text-primary mb-3 capitalize">${getCategoryTitle(category)}</h3>
-        <div class="space-y-2">
-          ${items.map((item, idx) => {
-            const key = `${category}-${idx}`;
-            const checked = this.progress[key] || false;
-            return `<label class="checklist-item flex items-start space-x-2 cursor-pointer ${checked ? 'completed' : ''}">
-              <input type="checkbox" class="mt-1 rounded border-gray-300 text-blue-600 focus:ring-blue-500" onchange="checklistComponent.toggleItem(this, '${key}')" ${checked ? 'checked' : ''}>
-              <span class="text-primary text-sm">${item}</span>
-            </label>`;
-          }).join('')}
-        </div>
-        <div class="mt-3 text-xs text-muted">
-          <span class="completed-count">${this.getCompletedCount(category)}</span> из ${items.length} выполнено
-        </div>`;
-      this.container.appendChild(div);
-    });
-  }
-
-  toggleItem(cb, key) {
-    const item = cb.closest('.checklist-item');
-    const counter = item.closest('.bg-white').querySelector('.completed-count');
-    if (cb.checked) {
-      item.classList.add('completed');
-      this.progress[key] = true;
-    } else {
-      item.classList.remove('completed');
-      this.progress[key] = false;
-    }
-    const category = key.split('-')[0];
-    counter.textContent = this.getCompletedCount(category);
-    StorageService.saveChecklist(this.progress);
-  }
-
-  getCompletedCount(category) {
-    let count = 0;
-    Object.keys(this.progress).forEach(k => {
-      if (k.startsWith(category + '-') && this.progress[k]) count++;
-    });
-    return count;
-  }
-}
-
-if (typeof module !== 'undefined') { module.exports = { Checklist }; }
-
-
-class Budget {
-  constructor(data) {
-    this.data = data;
-    this.container = document.querySelector('#budget .grid');
-  }
-
-  render() {
-    const container = this.container;
-    container.innerHTML = '';
-
-    const euroAmounts = [
-      '1 570–2 620',
-      '440–610',
-      '175–305',
-      '90–175',
-      '525–875',
-      '265–435',
-      '175–435',
-    ];
-    let totalMin = 0, totalMax = 0;
-    euroAmounts.forEach(a => {
-      const [min,max] = a.split('–').map(v=>parseInt(v.replace(/\s/g,'')));
-      totalMin += min;
-      totalMax += max || min;
-    });
-
-    this.data.forEach(item => {
-      const el = document.createElement('div');
-      el.className = 'bg-white border border-gray-200 rounded-lg p-4';
-      el.innerHTML = `<h3 class="font-semibold text-primary mb-2">${item.category}</h3><p class="text-2xl font-bold text-primary mb-1">${item.amount}</p><p class="text-muted text-sm">${item.note}</p>`;
-      container.appendChild(el);
-    });
-
-    const totalDiv = document.createElement('div');
-    totalDiv.className = 'bg-blue-50 border border-blue-200 rounded-lg p-4 md:col-span-2';
-    totalDiv.innerHTML = `<h3 class="font-semibold text-primary mb-2">💰 Общий бюджет</h3><p class="text-3xl font-bold text-blue-700 mb-1">${totalMin.toLocaleString()}–${totalMax.toLocaleString()} €</p>`;
-    container.appendChild(totalDiv);
-  }
-}
-
-if (typeof module !== 'undefined') { module.exports = { Budget }; }
-
-
-class Notes {
-  constructor(notes, storage) {
-    this.notes = notes;
-    this.storage = storage;
-    this.container = document.getElementById('notes-container');
-  }
-
-  render() {
-    const container = this.container;
-    container.innerHTML = '';
-    this.notes.forEach((note, index) => {
-      const div = document.createElement('div');
-      div.className = 'note-bubble p-4';
-      div.innerHTML = `
-      <div class="flex justify-between items-start mb-2">
-        <span class="font-medium text-primary">${note.author}</span>
-        <span class="text-muted text-sm">${new Date(note.timestamp).toLocaleDateString()}</span>
-      </div>
-      <p class="text-primary">${note.text}</p>
-      <button onclick="notesComponent.deleteNote(${index})" class="mt-2 text-red-600 hover:text-red-800 text-sm">Удалить</button>`;
-      container.appendChild(div);
-    });
-  }
-
-  addNote() {
-    const input = document.getElementById('note-input');
-    const text = input.value.trim();
-    if (text) {
-      this.notes.unshift({ text, author: 'Участник', timestamp: new Date() });
-      input.value = '';
-      this.render();
-      this.storage.saveNotes(this.notes);
-    }
-  }
-
-  addLocationNote(day, location) {
-    const noteText = prompt(`Добавить заметку к "${location}":`);
-    if (noteText && noteText.trim()) {
-      this.notes.unshift({ text: `📍 ${location} (День ${day}): ${noteText.trim()}`, author: 'Участник', timestamp: new Date() });
-      this.render();
-      this.storage.saveNotes(this.notes);
-      modalComponent.close();
-    }
-  }
-
-  deleteNote(index) {
-    if (confirm('Удалить эту заметку?')) {
-      this.notes.splice(index, 1);
-      this.render();
-      this.storage.saveNotes(this.notes);
-    }
-  }
-}
-
-if (typeof module !== 'undefined') { module.exports = { Notes }; }
-
-
-// Initialize services and components
-const notesData = StorageService.loadNotes();
-const checklistProgress = StorageService.loadChecklist();
-
-const modalComponent = new Modal();
-const timelineComponent = new Timeline(tripData, modalComponent);
-const checklistComponent = new Checklist(checklistData, checklistProgress);
-const budgetComponent = new Budget(budgetData);
-const notesComponent = new Notes(notesData, StorageService);
-
-// DOMContentLoaded
-document.addEventListener('DOMContentLoaded', () => {
-  timelineComponent.render();
-  checklistComponent.render();
-  budgetComponent.render();
-  notesComponent.render();
-  timelineComponent.renderHiddenGems();
-  NavigationService.showSection(StorageService.loadCurrentSection());
-  document.body.classList.add('loaded');
-});
-
-// Close modal on Escape
-document.addEventListener('keydown', e => {
-  if (e.key === 'Escape') modalComponent.close();
-});
-
-// Outside click
-document.getElementById('location-modal').addEventListener('click', e => {
-  if (e.target === document.getElementById('location-modal')) modalComponent.close();
-});
-
-// PWA service worker
-if ('serviceWorker' in navigator) {
-  window.addEventListener('load', () => {
-    navigator.serviceWorker
-      .register('/sw.js')
-      .then(r => console.log('SW registered: ', r))
-      .catch(err => console.log('SW registration failed: ', err));
-  });
-}
-
-if (typeof module !== 'undefined') { module.exports = {}; }
